FlirtCraft-backend/requirements.txt
# Simple requirements for basic functionality
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis and background tasks
redis==5.0.1
rq==1.15.1

# Basic utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Authentication
PyJWT==2.8.0
supabase==2.0.3
httpx==0.24.1
gotrue==1.3.0